Update linux-config and build methodology
This patch adds feature that users can use a different folder than
tiny-linux itself to start the build process. So that we can start
from a proper position where the relationship between the Linux
kernel path and the initramfs path can be applied to the kernel
directory itself (i.e., users can build Linux using the linux-config
from the kernel src directory instead of the build directory).

My environment is a workspace folder, where there are the following
subfolders:
 linux busybox obj tiny-linux
The build process now can be started from:
 workspace> LINUX_DIR=linux BUSYBOX_DIR=busybox ./tiny-linux/scripts/build.sh
 workspace/tiny-linux> LINUX_DIR=../linux BUSYBOX_DIR=../buxybox ./script/build.sh
And we can also copy the default kernel config to the kernel source
directory:
 workspace/linux> cp ../tiny-linux/config/config-linux-xxx ./.config
 workspace/linux> make oldconfig; make

This patch also update the linux-config with the following
modifications:
1. Minimal required configurations (added goldfish-rtc though it may
   not be required driver).
2. Use the latest v5.8-rc1 upstream kernel.

Signed-off-by: Lv Zheng <zhenglv@hotmail.com>